refactor: handler 코드를 통합 - 기존 예외 별 handler를 하나의 클래스에 통합
