--- src/main/java/io/github/potatoy/wanted_preonboarding_challenge/error/handler/AbstractExceptionHandler.java
@@ -1,8 +1,0 @@
-package io.github.potatoy.wanted_preonboarding_challenge.error.handler;
-
-import io.github.potatoy.wanted_preonboarding_challenge.error.dto.ErrorResponse;
-
-public abstract class AbstractExceptionHandler<T extends Exception> {
-
-  public abstract ErrorResponse handleException(T exception);
-}
