--- src/main/java/io/github/potatoy/wanted_preonboarding_challenge/error/handler/ForbiddenExceptionHandler.java
@@ -1,19 +1,0 @@
-package io.github.potatoy.wanted_preonboarding_challenge.error.handler;
-
-import io.github.potatoy.wanted_preonboarding_challenge.error.dto.ErrorResponse;
-import io.github.potatoy.wanted_preonboarding_challenge.error.exception.ForbiddenException;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ForbiddenExceptionHandler extends AbstractExceptionHandler<ForbiddenException> {
-
-  @Override
-  @ExceptionHandler(ForbiddenException.class)
-  @ResponseStatus(HttpStatus.FORBIDDEN)
-  public ErrorResponse handleException(ForbiddenException exception) {
-    return new ErrorResponse(HttpStatus.FORBIDDEN.value(), exception.getMessage());
-  }
-}
