--- src/main/java/io/github/potatoy/wanted_preonboarding_challenge/error/handler/NotFoundExceptionHandler.java
@@ -1,19 +1,0 @@
-package io.github.potatoy.wanted_preonboarding_challenge.error.handler;
-
-import io.github.potatoy.wanted_preonboarding_challenge.error.dto.ErrorResponse;
-import io.github.potatoy.wanted_preonboarding_challenge.error.exception.NotFoundException;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class NotFoundExceptionHandler extends AbstractExceptionHandler<NotFoundException> {
-
-  @Override
-  @ExceptionHandler(NotFoundException.class)
-  @ResponseStatus(HttpStatus.NOT_FOUND)
-  public ErrorResponse handleException(NotFoundException exception) {
-    return new ErrorResponse(HttpStatus.NOT_FOUND.value(), exception.getMessage());
-  }
-}
